+++ package.json
@@ -5,6 +5,8 @@
   "proxy": "http://localhost:4000",
   "private": true,
   "dependencies": {
+    "@nivo/core": "^0.67.0",
+    "@nivo/line": "^0.67.0",
     "@testing-library/jest-dom": "^5.11.9",
     "@testing-library/react": "^11.2.3",
     "@testing-library/user-event": "^12.6.2",
@@ -12,7 +14,9 @@
     "bootstrap": "^4.6.0",
     "dotenv": "^8.2.0",
     "express": "^4.17.1",
+    "express-session": "^1.17.1",
     "mapbox-gl": "^2.0.1",
+    "moment": "^2.29.1",
     "node-sass-chokidar": "^1.5.0",
     "npm-run-all": "^4.1.5",
     "react": "^17.0.1",
