+++ src/styles/components-scss/visualsComponent-scss/graphs-scss/_deathsGraph.scss
@@ -1,0 +1,2 @@
+@import '../../../base-scss/mixins';
+@import '../../../base-scss/variables';
