+++ src/styles/components-scss/visualsComponent-scss/graphs-scss/_netPositiveGraph.scss
@@ -1,0 +1,8 @@
+@import '../../../base-scss/mixins';
+@import '../../../base-scss/variables';
+
+.netPositive-graph {
+  height: 100%;
+  width: 100%;
+  // margin: 50px;
+}
