+++ src/styles/main.scss
@@ -17,3 +17,6 @@
 @import '../styles/components-scss/visuals';
 @import '../styles/components-scss/visualsComponent-scss/graph';
 @import '../styles/components-scss/visualsComponent-scss/map';
+
+@import '../styles/components-scss/visualsComponent-scss/graphs-scss/netPositiveGraph';
+@import '../styles/components-scss/visualsComponent-scss/graphs-scss/deathsGraph';
